Feat: index build is now part of docker build
