+++ app.py
@@ -18,48 +18,15 @@ META_FILE = "meta.pkl"
 CHAT_MODEL = os.getenv("CHAT_MODEL", "meta-llama/Meta-Llama-3-8B-Instruct")
 prometheus_helper = PrometheusHelper()
 
-# --- Build / load index ---
-def build_index():
-    prometheus_helper.start_index_build_timer()
-    print("Building FAISS index...")
-    ds = load_dataset("olivierdehaene/xkcd", split="train")
-    model = SentenceTransformer("all-MiniLM-L6-v2")
-    texts = []
-    for ex in ds:
-        title = ex["title"] if ex["title"] else ""
-        transcript = ex["transcript"] if ex["transcript"] else ""
-        explanation = ex["explanation"] if "explanation" in ex and ex["explanation"] else ""
-        texts.append(f"{title} {transcript} {explanation}")
-
-    embeddings = model.encode(texts, convert_to_numpy=True, show_progress_bar=True)
-    dim = embeddings.shape[1]
-    index = faiss.IndexFlatL2(dim)
-    index.add(embeddings)
-    faiss.write_index(index, INDEX_FILE)
-
-    # Store just the metadata we need (pickle-friendly)
-    meta = [
-        {
-            "id": ex["id"],
-            "title": ex["title"],
-            "transcript": ex["transcript"],
-            "explanation": ex["explanation"] if "explanation" in ex else "",
-        }
-        for ex in ds
-    ]
-    with open(META_FILE, "wb") as f:
-        pickle.dump(meta, f)
-
-    prometheus_helper.stop_index_build_timer()
-    return index, meta
-
+# --- Load index ---
 def get_index():
     if os.path.exists(INDEX_FILE) and os.path.exists(META_FILE):
         print("Loading cached index...")
         with open(META_FILE, "rb") as f:
             return faiss.read_index(INDEX_FILE), pickle.load(f)
     else:
-        return build_index()
+        print("Index files not found, please run build_index.py first.")
+        exit(1)
     
 def get_id_from_string(str:str) -> str:
     id_start = str.index("[") +1
